Cambios en el proyecto tras el analisis de sonarqube. Ademas de implementacion de tests para las clases Controller de los comerciales y clientes

+++ src/main/java/backend/filemanagment/FileServiceImpl.java
@@ -67,6 +67,8 @@ public class FileServiceImpl implements FileService {
     @Transactional
     public void deleteFile(Meeting meeting, Long fileId) {
         File file = fileRepository.findById(fileId).orElse(null);
+        if(file == null)
+            return;
         meeting.removeFile(file);
         fileRepository.delete(file);
     }
